+++ sharding-proxy/src/main/java/io/shardingsphere/shardingproxy/backend/jdbc/connection/BackendTransactionManager.java
@@ -23,7 +23,7 @@ import io.shardingsphere.core.constant.transaction.TransactionType;
 import io.shardingsphere.transaction.spi.ShardingTransactionHandler;
 import io.shardingsphere.transaction.core.internal.context.ShardingTransactionContext;
 import io.shardingsphere.transaction.core.internal.context.XATransactionContext;
-import io.shardingsphere.transaction.core.api.ShardingTransactionHandlerRegistry;
+import io.shardingsphere.transaction.core.loader.ShardingTransactionHandlerRegistry;
 import lombok.RequiredArgsConstructor;
 
 import java.sql.SQLException;
